fix (and add) autoload functionality
Instead of the solution in PR #8 of requiring multiple packages in
calibredb-utils.el, I have moved the single function calibredb to calibredb.el.
Besides that this fixes (recursive) autoload issues, this also fixes that that
the :config block when defered loading of calibredb with use-package is
evaluated (this is required for evilifying keybindings in the spacemacs layer).
I presumed that defining the function calibredb directly in calibredb.el does
not cause too much confusion.

Finally I made calibredb-find-counsel autoloadable for which some
packages are required from calibredb-ivy.el.