Update define-transient-command to define-transient-prefix and clean
The docstring of the define-transient-command notes that that function is
depcrecated and should be replaced with define-transient-prefix.

Furthermore, I found that there was a lot of code duplication for the
read-metadata functions. Instead I have created a 'master'
`calibredb-transient-read-metadata` that takes an argument for which type of
metadata should be called. I would argue this is a little cleaner.
(In the end I would like to add a function for the identifier field also,
because now the original value of that field is cleared when setting its
metadata. But for that I think the SQL query has to be extended.)