calibredb-query: use call-process-shell-command for error handling.
The command dump the output to a hidden buffer
called *calibredb-query-output*, then if it terminates successfully,
it will return the string of the output buffer. Else it switches to
the output buffer and tell user something’s wrong.

Also in calibredb-query-to-alist: Stop checking if the query-result is
nil because in that case, there is probably an error in
calibredb-query and calibredb-query will handle the error and halt
accordingly.